util.sh: ménage uniquement en cas de réussite
+ Si une compil échoue, on ne supprime pas son dossier de compilation: peut-être révélera-t-il des éléments intéressants post-mortem.

darcs-hash:c7d7b3a9aa6d917d25327cc40efe893374ddabd4

+++ util.sh
@@ -33,12 +33,15 @@ INSTALL_SCRIPTS="$SCRIPTS" # Des fois que d'autres récupèrent ensuite la varia
 
 util_menage()
 {
+	if [ $? -eq 0 ] # En cas de meurtre, on ne fait pas disparaître les preuves.
+	then
 	# Un minimum de blindage pour éviter de supprimer / en cas de gros, gros problème (genre le shell ne saurait même plus fournir $$).
 	case "$TMP/$$" in
 		*/[0-9]*)
 			rm -Rf "$TMP/$$"
 			;;
 	esac
+	fi
 }
 util_mechantmenage()
 {
